feat: rename externalDb.js of the API
