+++ api/src/neo4j/index.js
@@ -25,7 +25,7 @@ import {
   getComparisonOverview,
   getComparisonDetails,
 } from 'neo4j/queries/compare';
-import getComponentsForExternalDb from 'neo4j/queries/externalDb';
+import getComponentsForExternalDb from 'neo4j/queries/idInModels';
 
 export {
   getCompartment,
